Capstone Project: Facial Keypoint Detection (CV2 + CNN)
This is the Computer Vision capstone project in the Udacity AI
Nanodegree program. The programmed code is able to take in any image (or
video from web camera) containing faces and identify the location of
each face and their facial keypoints, hide identity, even put sunglusses
on the detected faces!